+++ src/views/AboutView.vue
@@ -1,6 +1,7 @@
 <template>
   <div class="about">
     <h1>This is an about page</h1>
+    <p>This is all about the blockchain admin panel.</p>
   </div>
 </template>