+++ src/api/http-commons.ts
@@ -1,7 +1,6 @@
 import axios from "axios";
 
 export const HTTP = axios.create({
-    // baseURL: "https://ninehearts-sandbox.com/api/v1",
     baseURL: "http://localhost:8080/api/v1",
     headers: {
         "Content-type": "application/json"