+++ src/components/HealthChart.vue
@@ -1,60 +1,46 @@
 <template>
   <div class="meter">
-    <h3>healthy: {{ healthAverage }}</h3>
+    <h3>healthy: {{ (loading) ? 'loading' : healthAverage }}</h3>
     <div>
       <canvas ref="healthChart"></canvas>
     </div>
   </div>
 </template>
 
-
 <script setup lang="ts">
-import {onMounted, ref} from "vue";
-import { useHealthStore } from "@/stores/health";
 import { Chart, registerables } from "chart.js";
 import type { ChartItem } from "chart.js";
-
-Chart.register(...registerables)
-
-const healthChart = ref(null)
+import { onMounted, ref } from "vue";
+import { useHealthStore } from "@/stores/health";
 
 const healthStore = useHealthStore()
+const healthChart = ref(null)
 const healthy = ref([0.0, 0.0, 0.0, 0.0, 0.0])
 const healthAverage = ref(0.0)
+const loading = ref(true)
+
+Chart.register(...registerables)
 
 const setHealth = () => {
   healthy.value = healthStore.health
   healthAverage.value = healthStore.average
 }
 
-const setHealthChartData = () => {
-  // healthChartData.value.data.datasets[0].data[0] = healthy.value
-}
+setInterval(async () => {
+  await healthStore.checkHealth()
+  setHealth()
+}, 10000)
 
 onMounted(async () => {
-  // await useHealthStore().checkReadiness()
-  // setReady()
   await healthStore.checkHealth()
   setHealth()
-  // await getHealth()
-  // setHealthChartData()
   renderChart()
 })
-
-setInterval(async () => {
-  // setReady()
-  await healthStore.checkHealth()
-  setHealth()
-  // renderChart()
-  // myChart.update()
-  // setHealthChartData()
-}, 5000)
 
 const renderChart = () => {
   const canvas = <unknown> healthChart.value as ChartItem
 
   if (canvas) {
-
       new Chart(canvas, {
       type: "radar",
       data: {
@@ -62,7 +48,6 @@ const renderChart = () => {
         datasets: [{
           label: 'Health',
           data: [healthy.value[0], healthy.value[1], healthy.value[2], healthy.value[3], healthy.value[4]],
-
           backgroundColor: [
             'rgba(130, 130, 255, 0.2)'
           ],
@@ -81,8 +66,8 @@ const renderChart = () => {
       }
     })
   }
+  loading.value = false
 }
-
 </script>
 
 <style scoped>