+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Unreleased
 
 *v0.0.1*
 
-**9/XX/23**
+**10/22/23**
 
 - [ADD Dashboard for Overall Health Metrics](#32)